requirements.txt
# ── Core Data Science ─────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.0

# ── NLP ───────────────────────────────────────────────────────────────────────
nltk==3.8.1

# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.0.3
Flask-Cors==4.0.1

# ── Model Serialisation ───────────────────────────────────────────────────────
joblib==1.4.2

# ── Environment / Config ──────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Visualisation (Jupyter Notebook) ─────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2

# ── Jupyter ───────────────────────────────────────────────────────────────────
jupyter==1.0.0
notebook==7.2.0
ipykernel==6.29.4

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0
